Improved dockerfile
Without this patch, the image for building the repo contains
the latest repositories, and needs to be maintained.

This is a problem, as the image needs to be updated to have
the latest version of the packages, where in fact it's not really
necessary: We could have a lightweight image that's only
changed when the code of this repo changes, and _USES_ the latest
repositories at runtime.

This fixes it by using an entrypoint which automatically updates
the base repositories, and then builds the user repository with
our existing repo-builder.sh